refactor: move getAnalyticsEndpoint and getAdaptedVisualization
